+++ Sources/XcodeGen/main.swift
@@ -3,6 +3,6 @@ import ProjectSpec
 import XcodeGenCLI
 import Version
 
-let version = Version("2.13.0")
+let version = Version("2.13.1")
 let cli = XcodeGenCLI(version: version)
 cli.execute()