+++ CHANGELOG.md
@@ -5,6 +5,9 @@
 #### Fixed
 - Validate scheme test action and test coverage target references before generating. [#775](https://github.com/yonaskolb/XcodeGen/pull/775) @liamnichols
 
+#### Internal
+- Updated to XcodeProj 7.8.0 [#777](https://github.com/yonaskolb/XcodeGen/pull/777) @yonaskolb
+
 ## 2.13.0
 
 #### Added